HADOOP-1879.  Remove some unneeded casts.  Contributed by Nilay Vaish.
git-svn-id: https://svn.apache.org/repos/asf/lucene/hadoop/trunk@575455 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/dfs/DFSClient.java
@@ -685,7 +685,7 @@ class DFSClient implements FSConstants {
                               firstChunkOffset + " != " + chunkOffset);
       }
       
-      int chunkLen = (int) in.readInt();
+      int chunkLen = in.readInt();
       
       // Sanity check the lengths
       if ( chunkLen < 0 || chunkLen > bytesPerChecksum ||
