+++ src/java/org/apache/hadoop/io/WritableUtils.java
@@ -394,7 +394,7 @@ public final class WritableUtils  {
     int total = 0;
     int cur = 0;
 
-    while ((total<len) && ((cur = (int) in.skipBytes(len-total)) > 0)) {
+    while ((total<len) && ((cur = in.skipBytes(len-total)) > 0)) {
         total += cur;
     }
